+++ src/lexer/utils/index.ts
@@ -5,3 +5,4 @@ export * as javascript from "./javascript";
 export * as json from "./json";
 export * from "./numeric";
 export * from "./subaction";
+export * from "./rejecter";